HGZ-39 [➕] - Damageable Architecture
# 데미지를 받을 수 있는 구조 변경
  - Unit은 IDamageable를 상속받음.
  - IDamageable은 데미지를 받을 수 있음.

# 인터페이스
  - IDamage (데미지 정보)
  - IDamageable (데미지를 받을 수 있는)
  - ICombative (공격이 가능한)

+++ HowToSurviveInThatWorld/Assets/__Scripts__/Model/_Core_/StatusSystem/Modular/StatModifier.cs
@@ -7,11 +7,11 @@ using UnityEngine;
 /// </summary>
 public class StatModifier
 {
-    #region Field
+    #region Fields
 
     public Object Source { get; set; }
     // 얼만큼 수정이 될 것인지에 대한 밸류 값
-    public int Magnitude { get; set; }
+    public float Magnitude { get; set; }
     // Magnitude를 어떤 타입에 형태로 (더할 것인지 곱할 것인지 등)
     public E_StatModifier_OperatorType Type { get; set; }
